Add support for indenting tagged bindings
This patch implements and uses an alternate forward-sexp fn which will
skip over ^ prefixed forms considered metadata by the Clojure
reader. This allows align-cljlet to correctly indent forms such as

    (let [^Long x 3
          y 3]
      (+ x y))

to

    (let [^Long x 3
          y       3]
      (+ x y))

since the symbol `Long` is seen to be metadata on whatever the following
sexp is, and consequently the width of the binding is the sum of the
widths of the binding form and all its leading metadata modifiers.

As this same forward-sexp is used for binding value forms, metadata
there will also be ignored and so forms such as

    (let [x ^StringBuilder (make-stringbuilder)
          y 3]
      y)

will be understood to be correctly indented rather than thought to have
multiple pairs on a line and be invalid.